rename parameters.py to ridge.py
