+++ robustness_test/common_utils/ridge.py
@@ -1,5 +1,6 @@
 import himalaya
 import numpy as np
+from ridge_utils.utils import counter
 
 GROUP_CV_SOVER_PARAMS = dict(n_iter=1, n_targets_batch=100, n_targets_batch_refit=100, alphas=np.logspace(0, 4, 10),
                              score_func=himalaya.scoring.correlation_score, progress_bar=True)
